requirements.txt
langchain==1.1.3
fastapi==0.116.1
langchain-huggingface==1.1.0
langchain-community==0.4.1
langchain-ollama==1.0.0
langchain-core==1.1.3
pycomcigan==1.4.0
bs4==0.0.2
pydantic==2.10.6
pydantic_core==2.27.2
pycomcigan==1.4.0
sentence-transformers==5.2.0
chromadb==1.3.7
uvicorn==0.40.0
# tf_keras==2.20.1
langchain_chroma==1.1.0
google-genai==1.56.0
google-api_core==2.28.1 
proto-plus==1.27.0    